feature_extraction/semantic-icon-classifier/requirements.txt
absl-py==0.9.0
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
bleach==1.5.0
cloudpickle==1.1.1
cudamat==0.3
cycler==0.10.0
dask==1.2.2
dominate==2.3.5
easydict==1.9
enum34==1.1.9
funcsigs==1.0.2
futures==3.3.0
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
imageio==2.5.0
joblib==0.14.1
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
matplotlib==2.1.0
mock==2.0.0
networkx==2.2
numpy==1.16.1
opencv-python==4.1.0.25
pandas==0.24.2
pathlib==1.0.1
pbr==5.4.4
Pillow==5.2.0
protobuf==3.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.3
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.95
scikit-image==0.14.2
scikit-learn==0.20.4
scipy==1.2.0
six==1.14.0
subprocess32==3.5.4
tensorboard==1.8.0
tensorboardX==1.6
tensorflow==1.8.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toolz==0.9.0
torchfile==0.1.0
tqdm==4.32.1
UNKNOWN==4.0.1
Werkzeug==1.0.0
wrapt==1.12.0
